+++ rtsp-server/xop/G711ASource.cpp
@@ -59,7 +59,7 @@ bool G711ASource::HandleFrame(const MediaChannelId channel_id,
 	RtpPacket rtp_pkt;
 	rtp_pkt.type = frame.type;
 	rtp_pkt.timestamp = frame.timestamp;
-	rtp_pkt.size = frame_size + RTP_TCP_HEAD_SIZE + RTP_HEADER_SIZE;
+	rtp_pkt.size = static_cast<uint16_t>(frame_size) + RTP_TCP_HEAD_SIZE + RTP_HEADER_SIZE;
 	rtp_pkt.last = 1;
 
 	memcpy(rtp_pkt.data.get() + RTP_TCP_HEAD_SIZE + RTP_HEADER_SIZE,
